# Speed up function `_moments_raw_to_central_fast` by 48%