### 📄 48% (0.48x) speedup for ***`_moments_raw_to_central_fast` in `skimage/measure/_moments_analytical.py`***

⏱️ Runtime :   **`1.06 milliseconds`**  **→** **`716 microseconds`** (best of `107` runs)
### 📝 Explanation and details

Here's how you can **significantly speed up** `_moments_raw_to_central_fast` by.

- Reducing redundant work and allocations (`np.zeros_like` is expensive and often not needed for small moment tables).
- Directly writing the correct minimal sized output array.
- Caching repeated scalar calculations.
- Skipping unnecessary `.astype` and dtype branching (just cast once up-front if needed).
- Avoiding large temporary arrays or multiple copies.
- Moves the `astype` to the very beginning, removing dtype branching altogether.
- Uses in-place computation where possible.



### Key optimizations.
- **Inlined repeated indexing:** All frequently accessed elements are cached as variables.
- **Reduced unnecessary operations:** Removed needless casts and allocations.
- **Kept all outputs in small contiguous arrays.**
- **Kept input shape checks and error raising, but otherwise maximized hot path efficiency.**

**This will be significantly faster, especially on small arrays and in tight loops.**


✅ **Correctness verification report:**

| Test                        | Status            |
| --------------------------- | ----------------- |
| ⚙️ Existing Unit Tests | 🔘 **None Found** |
| 🌀 Generated Regression Tests | ✅ **53 Passed** |
| ⏪ Replay Tests | ✅ **72 Passed** |
| 🔎 Concolic Coverage Tests | 🔘 **None Found** |
|📊 Tests Coverage       | 100.0% |
<details>
<summary>🌀 Generated Regression Tests Details</summary>

```python
import numpy as np
# imports
import pytest  # used for our unit tests
from skimage.measure._moments_analytical import _moments_raw_to_central_fast

# -----------------------------
# Unit tests for _moments_raw_to_central_fast
# -----------------------------

# --- Basic Test Cases ---

def test_2d_order1_basic():
    # 2D, order 1: only m[0,0], m[1,0], m[0,1] are meaningful
    # Central moment of order 0 is always the same as raw
    arr = np.zeros((2, 2), dtype=float)
    arr[0, 0] = 10.0
    arr[1, 0] = 20.0
    arr[0, 1] = 30.0
    codeflash_output = _moments_raw_to_central_fast(arr); result = codeflash_output # 6.31μs -> 3.23μs

def test_2d_order2_basic():
    # 2D, order 2: test with simple values
    arr = np.zeros((3, 3), dtype=float)
    arr[0, 0] = 4.0
    arr[1, 0] = 6.0
    arr[0, 1] = 8.0
    arr[2, 0] = 15.0
    arr[0, 2] = 20.0
    arr[1, 1] = 12.0
    codeflash_output = _moments_raw_to_central_fast(arr); result = codeflash_output # 7.22μs -> 4.16μs
    cx = arr[1, 0] / arr[0, 0]
    cy = arr[0, 1] / arr[0, 0]

def test_2d_order3_basic():
    # 2D, order 3: all moments up to [3,0] and [0,3] are present
    arr = np.zeros((4, 4), dtype=float)
    arr[0, 0] = 2.0
    arr[1, 0] = 3.0
    arr[0, 1] = 4.0
    arr[2, 0] = 5.0
    arr[0, 2] = 6.0
    arr[1, 1] = 7.0
    arr[3, 0] = 8.0
    arr[0, 3] = 9.0
    arr[2, 1] = 10.0
    arr[1, 2] = 11.0
    codeflash_output = _moments_raw_to_central_fast(arr); result = codeflash_output # 11.6μs -> 6.70μs
    cx = arr[1, 0] / arr[0, 0]
    cy = arr[0, 1] / arr[0, 0]
    cx2 = cx * cx
    cy2 = cy * cy
    cxcy = cx * cy
    # Check a few 3rd order central moments
    expected_21 = (
        arr[2, 1]
        - 2 * cx * arr[1, 1]
        - cy * arr[2, 0]
        + cx2 * arr[0, 1]
        + cxcy * arr[1, 0]
    )
    expected_12 = (
        arr[1, 2] - 2 * cy * arr[1, 1] - cx * arr[0, 2] + 2 * cxcy * arr[0, 1]
    )
    expected_30 = arr[3, 0] - 3 * cx * arr[2, 0] + 2 * cx2 * arr[1, 0]
    expected_03 = arr[0, 3] - 3 * cy * arr[0, 2] + 2 * cy2 * arr[0, 1]

def test_3d_order1_basic():
    # 3D, order 1: only m[0,0,0], m[1,0,0], m[0,1,0], m[0,0,1]
    arr = np.zeros((2, 2, 2), dtype=float)
    arr[0, 0, 0] = 5.0
    arr[1, 0, 0] = 2.0
    arr[0, 1, 0] = 3.0
    arr[0, 0, 1] = 4.0
    codeflash_output = _moments_raw_to_central_fast(arr); result = codeflash_output # 6.94μs -> 3.53μs

def test_3d_order2_basic():
    # 3D, order 2: check a few central moments
    arr = np.zeros((3, 3, 3), dtype=float)
    arr[0, 0, 0] = 10.0
    arr[1, 0, 0] = 2.0
    arr[0, 1, 0] = 3.0
    arr[0, 0, 1] = 4.0
    arr[2, 0, 0] = 5.0
    arr[0, 2, 0] = 6.0
    arr[0, 0, 2] = 7.0
    arr[1, 1, 0] = 8.0
    arr[1, 0, 1] = 9.0
    arr[0, 1, 1] = 11.0
    codeflash_output = _moments_raw_to_central_fast(arr); result = codeflash_output # 8.91μs -> 5.52μs
    cx = arr[1, 0, 0] / arr[0, 0, 0]
    cy = arr[0, 1, 0] / arr[0, 0, 0]
    cz = arr[0, 0, 1] / arr[0, 0, 0]

def test_3d_order3_basic():
    # 3D, order 3: check a couple of third order moments
    arr = np.zeros((4, 4, 4), dtype=float)
    arr[0, 0, 0] = 3.0
    arr[1, 0, 0] = 2.0
    arr[0, 1, 0] = 4.0
    arr[0, 0, 1] = 5.0
    arr[2, 0, 0] = 6.0
    arr[0, 2, 0] = 7.0
    arr[0, 0, 2] = 8.0
    arr[3, 0, 0] = 9.0
    arr[0, 3, 0] = 10.0
    arr[0, 0, 3] = 11.0
    arr[1, 1, 0] = 12.0
    arr[1, 0, 1] = 13.0
    arr[0, 1, 1] = 14.0
    arr[2, 1, 0] = 15.0
    arr[1, 2, 0] = 16.0
    arr[1, 1, 1] = 17.0
    arr[2, 0, 1] = 18.0
    arr[1, 0, 2] = 19.0
    arr[0, 2, 1] = 20.0
    arr[0, 1, 2] = 21.0
    arr[2, 0, 1] = 22.0
    arr[0, 2, 1] = 23.0
    arr[0, 1, 2] = 24.0
    arr[2, 1, 0] = 25.0
    arr[1, 2, 0] = 26.0
    arr[3, 0, 0] = 27.0
    arr[0, 3, 0] = 28.0
    arr[0, 0, 3] = 29.0
    # Only check a couple of moments for correctness
    codeflash_output = _moments_raw_to_central_fast(arr); result = codeflash_output # 16.9μs -> 11.4μs
    cx = arr[1, 0, 0] / arr[0, 0, 0]
    cy = arr[0, 1, 0] / arr[0, 0, 0]
    cz = arr[0, 0, 1] / arr[0, 0, 0]
    cx2 = cx * cx
    cy2 = cy * cy
    cz2 = cz * cz
    # Check [3,0,0]
    expected_300 = 2 * cx2 * arr[1, 0, 0] - 3 * cx * arr[2, 0, 0] + arr[3, 0, 0]
    # Check [0,0,3]
    expected_003 = 2 * cz2 * arr[0, 0, 1] - 3 * cz * arr[0, 0, 2] + arr[0, 0, 3]

# --- Edge Test Cases ---





def test_invalid_ndim():
    # 1D (not 2D or 3D)
    arr = np.zeros((3,), dtype=float)
    with pytest.raises(ValueError):
        _moments_raw_to_central_fast(arr)
    # 4D (not supported)
    arr4 = np.zeros((2, 2, 2, 2), dtype=float)
    with pytest.raises(ValueError):
        _moments_raw_to_central_fast(arr4)

def test_invalid_order():
    # 2D, order 4 (not supported)
    arr = np.zeros((5, 5), dtype=float)
    with pytest.raises(ValueError):
        _moments_raw_to_central_fast(arr)
    # 3D, order 4 (not supported)
    arr = np.zeros((5, 5, 5), dtype=float)
    with pytest.raises(ValueError):
        _moments_raw_to_central_fast(arr)

def test_dtype_preservation():
    # Should preserve float32 dtype in output
    arr = np.zeros((3, 3), dtype=np.float32)
    arr[0, 0] = 1.0
    arr[1, 0] = 2.0
    arr[0, 1] = 3.0
    arr[2, 0] = 4.0
    arr[0, 2] = 5.0
    arr[1, 1] = 6.0
    codeflash_output = _moments_raw_to_central_fast(arr); result = codeflash_output # 15.2μs -> 12.7μs

def test_noncontiguous_input():
    # Should work with non-contiguous arrays (e.g., a slice)
    arr = np.zeros((3, 6), dtype=float)
    arr[0, 0] = 1.0
    arr[1, 0] = 2.0
    arr[0, 1] = 3.0
    arr[2, 0] = 4.0
    arr[0, 2] = 5.0
    arr[1, 1] = 6.0
    # Take every other column (non-contiguous)
    arr2 = arr[:, ::2]
    codeflash_output = _moments_raw_to_central_fast(arr2); result = codeflash_output # 8.18μs -> 4.65μs

def test_negative_and_large_values():
    # Test with negative and large values
    arr = np.zeros((3, 3), dtype=float)
    arr[0, 0] = 1e10
    arr[1, 0] = -1e9
    arr[0, 1] = 2e9
    arr[2, 0] = 1e8
    arr[0, 2] = -2e8
    arr[1, 1] = 5e7
    codeflash_output = _moments_raw_to_central_fast(arr); result = codeflash_output # 7.89μs -> 4.28μs

# --- Large Scale Test Cases ---

def test_large_2d_order3():
    # Large 2D, order 3 (shape 4x4)
    arr = np.random.rand(4, 4) * 1000
    arr[0, 0] += 1e4  # Ensure nonzero mass
    codeflash_output = _moments_raw_to_central_fast(arr); result = codeflash_output # 11.8μs -> 7.26μs

def test_large_3d_order3():
    # Large 3D, order 3 (shape 4x4x4)
    arr = np.random.rand(4, 4, 4) * 1000
    arr[0, 0, 0] += 1e4  # Ensure nonzero mass
    codeflash_output = _moments_raw_to_central_fast(arr); result = codeflash_output # 19.0μs -> 12.3μs

def test_performance_large_2d():
    # Test with a large 2D array, order 3, shape (4, 1000)
    arr = np.random.rand(4, 1000)
    arr[0, 0] += 1e4  # Ensure nonzero mass
    codeflash_output = _moments_raw_to_central_fast(arr); result = codeflash_output # 12.9μs -> 8.55μs

def test_performance_large_3d():
    # Test with a large 3D array, order 3, shape (4, 10, 10)
    arr = np.random.rand(4, 10, 10)
    arr[0, 0, 0] += 1e4  # Ensure nonzero mass
    codeflash_output = _moments_raw_to_central_fast(arr); result = codeflash_output # 17.2μs -> 12.0μs
# codeflash_output is used to check that the output of the original code is the same as that of the optimized code.

import numpy as np
# imports
import pytest  # used for our unit tests
from skimage.measure._moments_analytical import _moments_raw_to_central_fast

# unit tests

# -------------------- Basic Test Cases --------------------

def test_2d_order1_identity():
    # 2D, order 1: central moments should be the same as raw for [0,0]
    m = np.zeros((2, 2), dtype=float)
    m[0, 0] = 10.0
    m[1, 0] = 20.0
    m[0, 1] = 30.0
    codeflash_output = _moments_raw_to_central_fast(m); out = codeflash_output # 8.42μs -> 3.93μs

def test_2d_order2_basic():
    # 2D, order 2: check central moment formulas
    m = np.zeros((3, 3), dtype=float)
    m[0, 0] = 4.0
    m[1, 0] = 8.0   # mean x = 2.0
    m[0, 1] = 12.0  # mean y = 3.0
    m[2, 0] = 40.0
    m[0, 2] = 100.0
    m[1, 1] = 24.0
    codeflash_output = _moments_raw_to_central_fast(m); out = codeflash_output # 7.22μs -> 4.16μs
    cx, cy = 2.0, 3.0

def test_2d_order3_basic():
    # 2D, order 3: check central moment formulas
    m = np.zeros((4, 4), dtype=float)
    m[0, 0] = 2.0
    m[1, 0] = 4.0
    m[0, 1] = 6.0
    m[2, 0] = 20.0
    m[0, 2] = 54.0
    m[1, 1] = 12.0
    m[3, 0] = 120.0
    m[0, 3] = 162.0
    m[2, 1] = 40.0
    m[1, 2] = 36.0
    cx, cy = 2.0, 3.0
    codeflash_output = _moments_raw_to_central_fast(m); out = codeflash_output # 11.6μs -> 6.70μs
    # [3,0] = m[3,0] - 3*cx*m[2,0] + 2*cx^2*m[1,0]
    expected_30 = 120.0 - 3*2.0*20.0 + 2*4.0*4.0
    # [0,3] = m[0,3] - 3*cy*m[0,2] + 2*cy^2*m[0,1]
    expected_03 = 162.0 - 3*3.0*54.0 + 2*9.0*6.0
    # [2,1] = m[2,1] - 2*cx*m[1,1] - cy*m[2,0] + cx^2*m[0,1] + cx*cy*m[1,0]
    expected_21 = 40.0 - 2*2.0*12.0 - 3.0*20.0 + 4.0*6.0 + 2.0*3.0*4.0
    # [1,2] = m[1,2] - 2*cy*m[1,1] - cx*m[0,2] + 2*cx*cy*m[0,1]
    expected_12 = 36.0 - 2*3.0*12.0 - 2.0*54.0 + 2*2.0*3.0*6.0

def test_3d_order1_identity():
    # 3D, order 1: only [0,0,0] is filled
    m = np.zeros((2, 2, 2), dtype=float)
    m[0, 0, 0] = 5.0
    m[1, 0, 0] = 10.0
    m[0, 1, 0] = 15.0
    m[0, 0, 1] = 20.0
    codeflash_output = _moments_raw_to_central_fast(m); out = codeflash_output # 7.45μs -> 3.66μs

def test_3d_order2_basic():
    # 3D, order 2: test a few central moments
    m = np.zeros((3, 3, 3), dtype=float)
    m[0, 0, 0] = 2.0
    m[1, 0, 0] = 4.0
    m[0, 1, 0] = 6.0
    m[0, 0, 1] = 8.0
    m[2, 0, 0] = 20.0
    m[0, 2, 0] = 54.0
    m[0, 0, 2] = 100.0
    m[1, 1, 0] = 12.0
    m[1, 0, 1] = 16.0
    m[0, 1, 1] = 24.0
    cx, cy, cz = 2.0, 3.0, 4.0
    codeflash_output = _moments_raw_to_central_fast(m); out = codeflash_output # 8.91μs -> 5.52μs

def test_3d_order3_basic():
    # 3D, order 3: check a few central moment formulas
    m = np.zeros((4, 4, 4), dtype=float)
    m[0, 0, 0] = 1.0
    m[1, 0, 0] = 2.0
    m[0, 1, 0] = 3.0
    m[0, 0, 1] = 4.0
    m[2, 0, 0] = 8.0
    m[0, 2, 0] = 27.0
    m[0, 0, 2] = 64.0
    m[1, 1, 0] = 6.0
    m[1, 0, 1] = 8.0
    m[0, 1, 1] = 12.0
    m[3, 0, 0] = 32.0
    m[0, 3, 0] = 81.0
    m[0, 0, 3] = 256.0
    m[2, 1, 0] = 24.0
    m[1, 2, 0] = 18.0
    m[2, 0, 1] = 16.0
    m[1, 0, 2] = 32.0
    m[0, 2, 1] = 36.0
    m[0, 1, 2] = 48.0
    m[2, 0, 1] = 16.0
    m[1, 1, 1] = 24.0
    cx, cy, cz = 2.0, 3.0, 4.0
    codeflash_output = _moments_raw_to_central_fast(m); out = codeflash_output # 16.9μs -> 11.4μs
    # [3,0,0] = 2*cx^2*m[1,0,0] - 3*cx*m[2,0,0] + m[3,0,0]
    expected_300 = 2*4.0*2.0 - 3*2.0*8.0 + 32.0
    # [0,3,0] = 2*cy^2*m[0,1,0] - 3*cy*m[0,2,0] + m[0,3,0]
    expected_030 = 2*9.0*3.0 - 3*3.0*27.0 + 81.0
    # [0,0,3] = 2*cz^2*m[0,0,1] - 3*cz*m[0,0,2] + m[0,0,3]
    expected_003 = 2*16.0*4.0 - 3*4.0*64.0 + 256.0

# -------------------- Edge Test Cases --------------------



def test_unsupported_ndim():
    # 1D: unsupported
    m = np.zeros((2,), dtype=float)
    with pytest.raises(ValueError):
        _moments_raw_to_central_fast(m)
    # 4D: unsupported
    m = np.zeros((2, 2, 2, 2), dtype=float)
    with pytest.raises(ValueError):
        _moments_raw_to_central_fast(m)

def test_unsupported_order():
    # 2D, order 4: unsupported
    m = np.zeros((5, 5), dtype=float)
    with pytest.raises(ValueError):
        _moments_raw_to_central_fast(m)
    # 3D, order 4: unsupported
    m = np.zeros((5, 5, 5), dtype=float)
    with pytest.raises(ValueError):
        _moments_raw_to_central_fast(m)

def test_integer_input_dtype():
    # Input is integer, output should be float but match expected
    m = np.zeros((3, 3), dtype=int)
    m[0, 0] = 4
    m[1, 0] = 8
    m[0, 1] = 12
    m[2, 0] = 40
    m[0, 2] = 100
    m[1, 1] = 24
    codeflash_output = _moments_raw_to_central_fast(m); out = codeflash_output # 14.3μs -> 12.9μs

def test_non_contiguous_input():
    # Test with a non-contiguous array (slice)
    base = np.zeros((6, 6), dtype=float)
    m = base[1:4, 2:5]  # shape (3,3)
    m[0, 0] = 4.0
    m[1, 0] = 8.0
    m[0, 1] = 12.0
    m[2, 0] = 40.0
    m[0, 2] = 100.0
    m[1, 1] = 24.0
    codeflash_output = _moments_raw_to_central_fast(m); out = codeflash_output # 8.21μs -> 4.59μs

def test_float32_input():
    # Input is float32, output should be float32
    m = np.zeros((3, 3), dtype=np.float32)
    m[0, 0] = 4.0
    m[1, 0] = 8.0
    m[0, 1] = 12.0
    m[2, 0] = 40.0
    m[0, 2] = 100.0
    m[1, 1] = 24.0
    codeflash_output = _moments_raw_to_central_fast(m); out = codeflash_output # 10.7μs -> 7.59μs

def test_negative_values():
    # Negative moments (not physically meaningful, but test math)
    m = np.zeros((3, 3), dtype=float)
    m[0, 0] = 4.0
    m[1, 0] = -8.0
    m[0, 1] = -12.0
    m[2, 0] = 40.0
    m[0, 2] = 100.0
    m[1, 1] = 24.0
    codeflash_output = _moments_raw_to_central_fast(m); out = codeflash_output # 7.59μs -> 4.34μs
    cx, cy = -2.0, -3.0

# -------------------- Large Scale Test Cases --------------------

def test_2d_large_random():
    # Large 2D moments, order 3, random values
    rng = np.random.default_rng(42)
    m = rng.random((4, 4)) * 1000 + 10
    m[0, 0] = 1e5  # ensure nonzero mass
    # Should not raise and output shape should match
    codeflash_output = _moments_raw_to_central_fast(m); out = codeflash_output # 13.2μs -> 7.53μs

def test_3d_large_random():
    # Large 3D moments, order 3, random values
    rng = np.random.default_rng(123)
    m = rng.random((4, 4, 4)) * 1000 + 10
    m[0, 0, 0] = 2e5  # ensure nonzero mass
    codeflash_output = _moments_raw_to_central_fast(m); out = codeflash_output # 19.0μs -> 12.5μs

def test_2d_many_batches():
    # Test a batch of many 2D moment arrays
    for i in range(10):
        m = np.zeros((3, 3), dtype=float)
        m[0, 0] = 10 + i
        m[1, 0] = 2 * (10 + i)
        m[0, 1] = 3 * (10 + i)
        m[2, 0] = 4 * (10 + i)
        m[0, 2] = 5 * (10 + i)
        m[1, 1] = 6 * (10 + i)
        codeflash_output = _moments_raw_to_central_fast(m); out = codeflash_output # 2.53μs -> 1.60μs

def test_3d_many_batches():
    # Test a batch of many 3D moment arrays
    for i in range(10):
        m = np.zeros((3, 3, 3), dtype=float)
        m[0, 0, 0] = 10 + i
        m[1, 0, 0] = 2 * (10 + i)
        m[0, 1, 0] = 3 * (10 + i)
        m[0, 0, 1] = 4 * (10 + i)
        m[2, 0, 0] = 5 * (10 + i)
        m[0, 2, 0] = 6 * (10 + i)
        m[0, 0, 2] = 7 * (10 + i)
        m[1, 1, 0] = 8 * (10 + i)
        m[1, 0, 1] = 9 * (10 + i)
        m[0, 1, 1] = 10 * (10 + i)
        codeflash_output = _moments_raw_to_central_fast(m); out = codeflash_output # 3.67μs -> 2.62μs
# codeflash_output is used to check that the output of the original code is the same as that of the optimized code.
```

</details>


To edit these changes `git checkout codeflash/optimize-_moments_raw_to_central_fast-mbrobm4w` and push.

[![Codeflash](https://img.shields.io/badge/Optimized%20with-Codeflash-yellow?style=flat&color=%23ffc428&logo=data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iNDgwIiBoZWlnaHQ9ImF1dG8iIHZpZXdCb3g9IjAgMCA0ODAgMjgwIiBmaWxsPSJub25lIiB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciPgo8cGF0aCBmaWxsLXJ1bGU9ImV2ZW5vZGQiIGNsaXAtcnVsZT0iZXZlbm9kZCIgZD0iTTI4Ni43IDAuMzc4NDE4SDIwMS43NTFMNTAuOTAxIDE0OC45MTFIMTM1Ljg1MUwwLjk2MDkzOCAyODEuOTk5SDk1LjQzNTJMMjgyLjMyNCA4OS45NjE2SDE5Ni4zNDVMMjg2LjcgMC4zNzg0MThaIiBmaWxsPSIjRkZDMDQzIi8+CjxwYXRoIGZpbGwtcnVsZT0iZXZlbm9kZCIgY2xpcC1ydWxlPSJldmVub2RkIiBkPSJNMzExLjYwNyAwLjM3ODkwNkwyNTguNTc4IDU0Ljk1MjZIMzc5LjU2N0w0MzIuMzM5IDAuMzc4OTA2SDMxMS42MDdaIiBmaWxsPSIjMEIwQTBBIi8+CjxwYXRoIGZpbGwtcnVsZT0iZXZlbm9kZCIgY2xpcC1ydWxlPSJldmVub2RkIiBkPSJNMzA5LjU0NyA4OS45NjAxTDI1Ni41MTggMTQ0LjI3NkgzNzcuNTA2TDQzMC4wMjEgODkuNzAyNkgzMDkuNTQ3Vjg5Ljk2MDFaIiBmaWxsPSIjMEIwQTBBIi8+CjxwYXRoIGZpbGwtcnVsZT0iZXZlbm9kZCIgY2xpcC1ydWxlPSJldmVub2RkIiBkPSJNMjQyLjg3MyAxNjQuNjZMMTg5Ljg0NCAyMTkuMjM0SDMxMC44MzNMMzYzLjM0NyAxNjQuNjZIMjQyLjg3M1oiIGZpbGw9IiMwQjBBMEEiLz4KPC9zdmc+Cg==)](https://codeflash.ai)